Add Hive support to identify cloud table
A cloud table is defined as a remote data source stored on a high-bandwdith storage
 media which supports many parallel connections and capable of high concurrency
connection requests for data extraction. Current implementation only supports
Amazon S3 as it uses a scalable storage infrastrcture with a top
limit of 5500 GET/HEAD reqeusts per second per S3 bucket.

For an identified cloud table, ConnectorTableLayout is tagged a boolean flag as
a table layout and it propagates the identity for TableHandle.

This support is implemented with Hive connector and may be extended for any feasible cloud storage.